replaced unmaintained dependencies with defacto standards
htmlparser, soup-select and unirest were all very old modules that
are not maintained anymore. therefore we replaced unirest with
node-request and htmlparser/soup-select with cheerio.

this is also solving issues with deprecated node apis being used,
which prevented this module being used in newer versions of node
or/and things like react-native.